refactor(core): decompose agent_task_finalization god-file into submodules
agent_task_finalization.rs was 2118 lines (god_file). It already had a
submodule dir (backend, proof, schemas) but the main file mixed ~300
lines of types, ~500 lines of finalization logic, and a ~1290-line inline
test module.

- Extract the type definitions (report/outcome/intent/proof structs,
  enums, backend trait, PR-ref types) into types.rs, re-exported via
  pub use types::*.
- Extract the inline test module into tests.rs (same module path, no
  super:: adjustment needed).

The main file keeps imports, schema consts, submodule declarations, and
the finalization logic (finalize_pr, gate validation, publication intent,
report building). All resulting files are under the 1500-line god_file
threshold (main 527, types 302, tests 1285).

No logic changed; public API preserved (pub use types::* keeps every
type export). Clears the agent_task_finalization.rs god_file finding.
